move iterator.rs to execution module as pager.rs
