+++ docs/source/migration-guides/0.15-deserialization.md
@@ -285,7 +285,7 @@ let legacy_result: LegacyQueryResult = result.into_legacy_result()?;
 
 ```rust
 # extern crate scylla;
-# use scylla::transport::iterator::{QueryPager, LegacyRowIterator};
+# use scylla::execution::pager::{QueryPager, LegacyRowIterator};
 # use std::error::Error;
 # async fn check_only_compiles(pager: QueryPager) -> Result<(), Box<dyn Error>> {
 let pager: QueryPager = pager;
