+++ scylla/src/cluster/metadata.rs
@@ -2,11 +2,11 @@ use crate::cluster::host_filter::HostFilter;
 use crate::cluster::node::resolve_contact_points;
 use crate::connection::{Connection, ConnectionConfig, NodeConnectionPool, PoolConfig, PoolSize};
 use crate::deserialize::DeserializeOwnedRow;
+use crate::execution::pager::QueryPager;
 use crate::frame::response::event::Event;
 use crate::routing::Token;
 use crate::statement::query::Query;
 use crate::transport::errors::{DbError, NewSessionError, QueryError};
-use crate::transport::iterator::QueryPager;
 use crate::utils::parse::{ParseErrorCause, ParseResult, ParserState};
 
 use futures::future::{self, FutureExt};
@@ -976,7 +976,7 @@ where
 
     let fut = async move {
         let pager = make_keyspace_filtered_query_pager(conn, query_str, keyspaces_to_fetch).await?;
-        let stream: crate::transport::iterator::TypedRowStream<R> =
+        let stream: crate::execution::pager::TypedRowStream<R> =
             pager.rows_stream::<R>().map_err(convert_typecheck_error)?;
         Ok::<_, QueryError>(stream)
     };
